+++ Main/Features/TroopProgression/Models/TaomCharacterStatsModel.cs
@@ -1,0 +1,8 @@
+using TaleWorlds.CampaignSystem.GameComponents;
+
+namespace TAOM.Features.TroopProgression.Models;
+
+public class TaomCharacterStatsModel : DefaultCharacterStatsModel
+{
+    public override int MaxCharacterTier => 10;
+}