+++ Main/Features/TroopProgression/VolunteerTierService.cs
@@ -1,0 +1,6 @@
+namespace TAOM.Features.TroopProgression;
+
+public class VolunteerTierService : IVolunteerTierService
+{
+    public int MaxVolunteerTier => 6;
+}